started reorg of metrics
